# Application Programming Interface/BuildingBug.API/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "BuildingBugAPI",
    "version": "v1"
  },
  "paths": {
...
    "/api/v{version}/Profissional/new": {
      "post": {
        "tags": [
          "Profissional"
        ],
        "summary": "Criar um Profissional",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateProfissionalRequestContract"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/v{version}/Profissional/all": {
      "get": {
        "tags": [
          "Profissional"
        ],
        "summary": "Obter todos os profissionais",
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/v{version}/Tipo/new": {
      "post": {
        "tags": [
          "Tipo"
        ],
        "summary": "Criar Tipo",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateTipoRequestContract"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/v{version}/Tipo/all": {
      "get": {
        "tags": [
          "Tipo"
        ],
        "summary": "Obter todos os Tipos",
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/v{version}/TipoProfissional/new": {
      "post": {
        "tags": [
          "TipoProfissional"
        ],
        "summary": "Criar tipo de profissional",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateTipoProfissionalRequestContract"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/TipoProfissional/delete": {
      "delete": {
        "tags": [
          "TipoProfissional"
        ],
        "summary": "Eliminar tipo de profissional",
        "parameters": [
          {
            "in": "query",
            "name": "tipoId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "profissionalId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/TipoProfissional/{tipoId}/{profissionalId}": {
      "get": {
        "tags": [
          "TipoProfissional"
        ],
        "summary": "Obter tipo de profissional",
        "parameters": [
          {
            "in": "path",
            "name": "tipoId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "profissionalId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/TipoProfissional/all": {
      "get": {
        "tags": [
          "TipoProfissional"
        ],
        "summary": "Obter todos os tipos de profissional",
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
...
    "CreateProfissionalRequestContract": {
      "required": [
        "apelido",
        "contacto",
        "email",
        "nome",
        "senha"
      ],
      "type": "object",
      "properties": {
        "nome": {
          "minLength": 1,
          "type": "string"
        },
        "apelido": {
          "minLength": 1,
          "type": "string"
        },
        "email": {
          "minLength": 1,
          "type": "string"
        },
        "senha": {
          "minLength": 1,
          "type": "string"
        },
        "contacto": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "CreateTipoProfissionalRequestContract": {
      "type": "object",
      "properties": {
        "tipo_id": {
          "format": "int32",
          "type": "integer"
        },
        "profissional_id": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "CreateTipoRequestContract": {
      "type": "object",
      "properties": {
        "descricao": {
          "type": "string"
        },
        "preco": {
          "format": "int32",
          "type": "integer"
        },
        "fator": {
          "type": "boolean"
        },
        "categoria": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Insert Token as: 'Bearer your_token_here'"
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}